+++ Source/Utils/UndoActions/RegionLifecycleUndoActions.h
@@ -140,7 +140,7 @@ public:
         return true;
     }
 
-    int getSizeInUnits() override { return sizeof(*this) + m_regions.size() * sizeof(Region); }
+    int getSizeInUnits() override { return static_cast<int>(sizeof(*this) + static_cast<size_t>(m_regions.size()) * sizeof(Region)); }
 
 private:
     RegionManager& m_regionManager;
@@ -292,7 +292,7 @@ public:
         return true;
     }
 
-    int getSizeInUnits() override { return sizeof(*this) + m_oldName.length() + m_newName.length(); }
+    int getSizeInUnits() override { return static_cast<int>(sizeof(*this)) + m_oldName.length() + m_newName.length(); }
 
 private:
     RegionManager& m_regionManager;
